+++ src/software/thunderscope/robot_diagnostics/motor_fault_view.py
@@ -6,7 +6,8 @@ from software.thunderscope.robot_diagnostics.motor_fault_icons.motor_fault_icon_
     get_stopped_icon,
     get_warning_icon,
 )
-from software.thunderscope.common.common_widgets import get_string_val, display_tooltip
+from software.thunderscope.common.proto_parameter_tree_util import get_string_val
+from software.thunderscope.common.common_widgets import display_tooltip
 
 
 class MotorFaultView(QWidget):
